requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.2
psycopg2-binary==2.9.1
flask-cors==3.0.10
Werkzeug==2.2.3